sub-process

+++ vm_monitor_agent/main.py
@@ -4,7 +4,6 @@ import api_server
 import agent
 from data_collector import wq
 from report_server_api import reporter
-from vm_monitor_agent.agent_rabbit import main as agent_rabbit_main
 
 if settings['cmd_way'] == 'meta_serv':
     from report_server_api import meta_checker as mc
@@ -13,7 +12,29 @@ import sys
 import os
 import logging
 import signal
-from multiprocessing import Process
+from subprocess import Popen
+
+# configure log
+import logging
+from os import path, makedirs
+
+log_level = {
+    'DEBUG': logging.DEBUG,
+    'INFO': logging.INFO,
+    'WARNING': logging.WARNING,
+    'ERROR': logging.ERROR,
+    'CRITICAL': logging.CRITICAL
+}
+
+log_dir = settings['log_file'].rpartition(settings['log_file'][settings['log_file'].rindex('/')])[0]
+if not path.exists(log_dir):
+    makedirs(log_dir)
+
+logging.basicConfig(level=log_level[settings['log_level']], \
+        format='[%(asctime)s] %(filename)s[line:%(lineno)d]: %(message)s', \
+        datefmt='%a, %d %b %Y %H:%M:%S', \
+        filename=settings['log_file'], \
+        filemode='a')
 
 LOG = logging.getLogger(__name__)
 
@@ -34,8 +55,9 @@ def exit_handler(signum, frame):
     wq.join()
 
     global Agent_Rabbit_Process
-    # os.kill(Agent_Rabbit_PID, signal.SIGTERM)
-    Agent_Rabbit_Process.terminate()
+    if Agent_Rabbit_Process is not None:
+        # os.kill(Agent_Rabbit_PID, signal.SIGTERM)
+        Agent_Rabbit_Process.terminate()
 
     sys.exit(0)
 
@@ -50,19 +72,19 @@ def main():
         LOG.info("Agent PID: %d" % os.getpid())
     except ValueError, e:
         LOG.warning(e)
-    
-    # start agent
-    agent.start()
-    wq.start()
-    reporter.start()
-    if settings['cmd_way'] == 'meta_serv':
-        mc.start()
 
     # start agent_rabbit
-    p = Process(target=agent_rabbit_main, name="AgentRabbit", \
-            args=(settings['instance_id'], ))
-    p.start()
+    LOG.info('VM UUID: %s' % settings['instance_id'])
+    p = Popen(['agent_rabbit', '-u', settings['instance_id']], shell=True)
     Agent_Rabbit_Process = p
+
+    if settings['start_proc_monitor']:
+        # start agent
+        agent.start()
+        wq.start()
+        reporter.start()
+        if settings['cmd_way'] == 'meta_serv':
+            mc.start()
 
     # start api server
     if settings['cmd_way'] == 'rest':
@@ -70,7 +92,7 @@ def main():
         sys.argv[1] = settings['port']
         api_server.run()
 
-    p.join()
+    p.wait()
 
 if __name__ == '__main__':
     main()
